+++ agentlace/trainer.py
@@ -5,10 +5,10 @@ from typing import Any, Optional, Callable, Set, Dict, List, Tuple, Any
 from typing_extensions import Protocol
 from collections import deque
 
-from edgeml.zmq_wrapper.req_rep import ReqRepServer, ReqRepClient
-from edgeml.zmq_wrapper.broadcast import BroadcastServer, BroadcastClient
-from edgeml.internal.utils import compute_hash
-from edgeml.data.data_store import DataStoreBase
+from agentlace.zmq_wrapper.req_rep import ReqRepServer, ReqRepClient
+from agentlace.zmq_wrapper.broadcast import BroadcastServer, BroadcastClient
+from agentlace.internal.utils import compute_hash
+from agentlace.data.data_store import DataStoreBase
 
 import time
 import threading
